+++ KlockWorkiOS/Pages/AppSettings.swift
@@ -1,0 +1,15 @@
+//
+//  AppSettings.swift
+//  KlockWorkiOS
+//
+//  Created by Ryan Priebe on 2024-05-22.
+//  Copyright © 2024 YegCollective. All rights reserved.
+//
+
+import SwiftUI
+
+struct AppSettings: View {
+    var body: some View {
+        Text("Coming soon!")
+    }
+}
