+++ KlockWorkiOS/Pages/Planning.swift
@@ -1,0 +1,15 @@
+//
+//  Planning.swift
+//  KlockWorkiOS
+//
+//  Created by Ryan Priebe on 2024-05-22.
+//  Copyright © 2024 YegCollective. All rights reserved.
+//
+
+import SwiftUI
+
+struct Planning: View {
+    var body: some View {
+        Text("Coming soon!")
+    }
+}
