+++ KlockWorkiOS/Pages/Today.swift
@@ -1,0 +1,15 @@
+//
+//  Today.swift
+//  KlockWorkiOS
+//
+//  Created by Ryan Priebe on 2024-05-22.
+//  Copyright © 2024 YegCollective. All rights reserved.
+//
+
+import SwiftUI
+
+struct Today: View {
+    var body: some View {
+        Text("Coming soon!")
+    }
+}
